refactor: Moved SVG rendering to `correlation_plot_data_utils.ts`
